clean: remove old classes
