+++ src/main/java/fr/inria/stamp/Ex2Amplifier.java
@@ -1,34 +1,28 @@
-package fr.inria.stamp;
+package fr.inria.stamp.catg;
 
 import fr.inria.diversify.automaticbuilder.AutomaticBuilderFactory;
 import fr.inria.diversify.dspot.amplifier.Amplifier;
+import fr.inria.diversify.dspot.assertGenerator.AssertionRemover;
 import fr.inria.diversify.dspot.support.DSpotCompiler;
 import fr.inria.diversify.utils.AmplificationHelper;
 import fr.inria.diversify.utils.DSpotUtils;
 import fr.inria.diversify.utils.sosiefier.InputConfiguration;
-import fr.inria.stamp.jbse.JBSERunner;
-import fr.inria.stamp.smt.SMTSolver;
 import spoon.reflect.code.CtLiteral;
-import spoon.reflect.code.CtUnaryOperator;
 import spoon.reflect.declaration.CtClass;
 import spoon.reflect.declaration.CtMethod;
-import spoon.reflect.declaration.CtParameter;
 import spoon.reflect.declaration.CtType;
-import spoon.reflect.reference.CtTypeReference;
+import spoon.reflect.factory.Factory;
 import spoon.reflect.visitor.filter.TypeFilter;
 
 import java.io.File;
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 /**
  * Created by Benjamin DANGLOT
  * benjamin.danglot@inria.fr
- * on 09/01/18
+ * on 24/01/18
  */
 public class Ex2Amplifier implements Amplifier {
 
@@ -40,67 +34,57 @@ public class Ex2Amplifier implements Amplifier {
 
     @Override
     public List<CtMethod> apply(CtMethod ctMethod) {
-        if (!canBeRun.test(ctMethod)) {
-            return Collections.emptyList();
-        }
-        final CtMethod<?> extractedMethod = ArgumentsExtractor.performExtraction(ctMethod);
-        final CtType ctType = ctMethod.getParent(CtClass.class);
-        CtType<?> clone = ctType.clone();
-        clone.setParent(ctType.getParent());
-        clone.removeMethod(ctMethod);
-        clone.addMethod(extractedMethod);
-        DSpotUtils.printJavaFileWithComment(clone, new File(DSpotCompiler.pathToTmpTestSources));
-        final String classpath = AutomaticBuilderFactory
+        final CtMethod<?> mainMethodFromTestMethod =
+                MainGenerator.generateMainMethodFromTestMethod(ctMethod);
+        final CtClass testClass = ctMethod.getParent(CtClass.class);
+        testClass.addMethod(mainMethodFromTestMethod);
+        DSpotUtils.printJavaFileWithComment(testClass, new File("target/dspot/tmp_test_sources"));
+        String classpath = AutomaticBuilderFactory
                 .getAutomaticBuilder(this.configuration)
                 .buildClasspath(this.configuration.getInputProgram().getProgramDir())
                 + AmplificationHelper.PATH_SEPARATOR +
                 this.configuration.getInputProgram().getProgramDir() + "/" + this.configuration.getInputProgram().getClassesDir()
-                + AmplificationHelper.PATH_SEPARATOR + "target/dspot/dependencies/"
                 + AmplificationHelper.PATH_SEPARATOR +
                 this.configuration.getInputProgram().getProgramDir() + "/" + this.configuration.getInputProgram().getTestClassesDir();
-        DSpotCompiler.compile(DSpotCompiler.pathToTmpTestSources, classpath,
-                new File(this.configuration.getInputProgram().getProgramDir() + "/" + this.configuration.getInputProgram().getTestClassesDir()));
-        final List<Map<String, List<String>>> conditionForEachParameterForEachState = JBSERunner.runJBSE(classpath, extractedMethod);
-        return conditionForEachParameterForEachState.stream()
-                .map(conditions ->
-                        this.generateNewTestMethod(ctMethod, conditions)
-                )
+        DSpotCompiler.compile("target/dspot/tmp_test_sources", classpath + AmplificationHelper.PATH_SEPARATOR + "lib/catg-dev.jar",
+                new File(this.configuration.getInputProgram().getProgramDir() + "/" +
+                        this.configuration.getInputProgram().getTestClassesDir())
+        );
+        final List<List<String>> execute = CATGExecutor.execute(classpath, testClass.getQualifiedName());
+        CATGUtils.eraseOldFiles();
+        testClass.removeMethod(mainMethodFromTestMethod);
+        return execute.stream()
+                .map(values -> buildMethodFromValues(values, ctMethod))
                 .collect(Collectors.toList());
     }
 
-    private CtMethod<?> generateNewTestMethod(CtMethod<?> testMethod,
-                                              Map<String, List<String>> conditionForParameter) {
-        final CtMethod clone = AmplificationHelper.cloneMethodTest(testMethod, "_examplifier");
-        final List<?> solutions = SMTSolver.solve(conditionForParameter);
-        final Iterator<?> iterator = solutions.iterator();
-        final List<CtLiteral> originalLiterals =
-                clone.getElements(new TypeFilter<>(CtLiteral.class));
-        conditionForParameter.keySet()
-                .forEach(s -> {
-                    final int indexOfLit = Integer.parseInt(s.substring("param".length()));
-                    final CtLiteral literalToBeReplaced = originalLiterals.get(indexOfLit);
-                    final CtLiteral<?> newLiteral = testMethod.getFactory().createLiteral(iterator.next());
-                    if (literalToBeReplaced.getParent() instanceof CtUnaryOperator) {
-                        literalToBeReplaced.getParent().replace(newLiteral);
-                    } else {
-                        literalToBeReplaced.replace(newLiteral);
-                    }
-                });
+    private CtMethod<?> buildMethodFromValues(List<String> values, CtMethod originalTestMethod) {
+        final Iterator<String> iteratorOnNewValues = values.iterator();
+        final CtMethod<?> clone = new AssertionRemover().removeAssertion(
+                AmplificationHelper.cloneMethodTest(originalTestMethod, "Ex2Amplifier")
+        );
+        final List<CtLiteral> originalLiterals = clone.getBody().getElements(new TypeFilter<CtLiteral>(CtLiteral.class));
+        originalLiterals.forEach(ctLiteral ->
+                    ctLiteral.replace(buildNewLiteralFromString(iteratorOnNewValues.next(), ctLiteral))
+                );
         return clone;
     }
 
-    private final Predicate<CtMethod<?>> canBeRun = ctMethod ->
-            ctMethod.getParameters()
-                    .stream()
-                    .map(CtParameter::getType)
-                    .map(CtTypeReference::getSimpleName)
-                    .allMatch(JBSERunner.typeToDescriptor::containsKey
-                            // || "String".equals(typeStr)  TODO for now, we do not support String, because JBSE might take long time
-                    );
+    private CtLiteral<?> buildNewLiteralFromString(String value, CtLiteral<?> originalLiteral) {
+        final Object originalLiteralValue = originalLiteral.getValue();
+        final Factory factory = originalLiteral.getFactory();
+        if (originalLiteralValue instanceof String) {
+            return factory.createLiteral(value);
+        } else if (originalLiteralValue instanceof Integer){
+            return factory.createLiteral(Integer.parseInt(value));
+        } else {
+            throw new UnsupportedOperationException(originalLiteralValue.getClass()  + " is not supported");
+        }
+    }
 
     @Override
     public CtMethod applyRandom(CtMethod ctMethod) {
-        return null;
+        throw new UnsupportedOperationException();
     }
 
     @Override
