+++ src/main/java/fr/inria/stamp/MainGenerator.java
