+++ src/test/java/fr/inria/stamp/MainGeneratorTest.java
