--- src/test/java/fr/inria/stamp/catg/Ex2AmplifierTest.java
@@ -1,41 +1,0 @@
-package fr.inria.stamp.catg;
-
-import fr.inria.stamp.AbstractTest;
-import org.junit.Test;
-import spoon.reflect.declaration.CtMethod;
-
-import java.util.List;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Created by Benjamin DANGLOT
- * benjamin.danglot@inria.fr
- * on 24/01/18
- */
-public class Ex2AmplifierTest extends AbstractTest {
-
-    @Override
-    protected String getPathToConfigurationFile() {
-        return "src/test/resources/tavern/tavern.properties";
-    }
-
-    @Test
-    public void test() throws Exception {
-
-        /*
-            Test that the Ex2Amplifier returns a List of CtMethod build thank to CATG.
-            Amplified CtMethod has the same "structural" inputs of the test, but with
-            different test data input, i.e. literals has been modified.
-         */
-
-        final Ex2Amplifier ex2Amplifier = new Ex2Amplifier(this.configuration);
-        final CtMethod<?> test = this.launcher.getFactory()
-                .Class()
-                .get("fr.inria.stamp.MainTest")
-                .getMethodsByName("test")
-                .get(0);
-        final List<CtMethod> apply = ex2Amplifier.apply(test);
-        assertEquals(3, apply.size());
-    }
-}
